Refactor Trigger Apex Test core implementation
Moved conversion of source to MDAPI format to external utility function.
